# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AICSS Backend",
    "description": "AI Cinematic Spatial System — Depth + Segmentation + Spatial Layers",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/aicss/v2/scripts/motion/generate": {
      "post": {
        "tags": [
          "v2 - Script & Motion",
          "Script & Motion"
        ],
        "summary": "Api Generate Motion",
        "description": "Full motion sequence pipeline.\n\nSteps:\n  1. Generate action video via DashScope (or other provider)\n  2. Extract PNG frames via ffmpeg\n  3. Segment person from frames via SAM2 (if available)\n\nPersists the resulting MotionSequence JSON to <project>/motions/<shot_id>.json\nwhen a project_id is provided.",
        "operationId": "api_generate_motion_api_aicss_v2_scripts_motion_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateMotionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MotionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenerateMotionRequest": {
        "properties": {
          "shot_id": {
            "type": "string",
            "title": "Shot Id"
          },
          "character_id": {
            "type": "string",
            "title": "Character Id"
          },
          "character_name": {
            "type": "string",
            "title": "Character Name"
          },
          "action_prompt": {
            "type": "string",
            "title": "Action Prompt"
          },
          "start_image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Image",
            "description": "Base64 start frame"
          },
          "end_image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Image",
            "description": "Base64 end frame"
          },
          "duration_seconds": {
            "type": "number",
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Duration Seconds",
            "default": 5.0
          },
          "video_provider": {
            "type": "string",
            "title": "Video Provider",
            "description": "Video provider: dashscope | local_wan | svd",
            "default": "dashscope"
          },
          "greenscreen": {
            "type": "boolean",
            "title": "Greenscreen",
            "description": "Render the action video against a flat green-screen background and apply a chroma-key pass during segmentation.",
            "default": false
          },
          "feather_edges": {
            "type": "boolean",
            "title": "Feather Edges",
            "description": "Snap SAM2 masks to nearby Canny edges to soften character silhouettes.",
            "default": true
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          }
        },
        "type": "object",
        "required": [
          "shot_id",
          "character_id",
          "character_name",
          "action_prompt"
        ],
        "title": "GenerateMotionRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MotionResponse": {
        "properties": {
          "shot_id": {
            "type": "string",
            "title": "Shot Id"
          },
          "character_id": {
            "type": "string",
            "title": "Character Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "video_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Path"
          },
          "frame_count": {
            "type": "integer",
            "title": "Frame Count",
            "default": 0
          },
          "segmented_frames": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Segmented Frames",
            "default": []
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          }
        },
        "type": "object",
        "required": [
          "shot_id",
          "character_id",
          "status"
        ],
        "title": "MotionResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}